fix: storage slider cap, install path failure, detection precision
- config.EnsureDirectories: skip os.MkdirAll("") when Policy.Directory
  is empty (embedded FS mode) — fixes "path not found" on Windows
- wizard/types.go: StorageInfo.TotalGB/AvailableGB and
  MemoryInfo.TotalGB/AvailableGB changed int → float64
- wizard/detection.go: float64 division for GB conversion; int() casts
  added in CalculateAvailableResources; format strings updated
- dashboard.go: slider max driven by state.Capabilities.Storage.AvailableGB;
  slider note shows actual free space; math.Floor for clean default;
  %.1f format throughout hardware display
- cost_calculator.go: removed now-redundant float64() casts
- dependencies.go, wizard_demo.go, main.go: %.1f GB format for all
  TotalGB/AvailableGB display sites

+++ internal/wizard/cost_calculator.go
@@ -51,7 +51,7 @@ func (c *CostCalculator) estimateIdlePower() float64 {
 	power += cpuTDP * 0.15
 
 	// RAM: ~3W per 8GB
-	ramPower := float64(c.capabilities.Memory.TotalGB) / 8.0 * 3.0
+	ramPower := c.capabilities.Memory.TotalGB / 8.0 * 3.0
 	power += ramPower
 
 	// Storage: 2-5W per drive
@@ -80,7 +80,7 @@ func (c *CostCalculator) estimateLoadPower() float64 {
 	power += cpuTDP
 
 	// RAM: ~3W per 8GB
-	ramPower := float64(c.capabilities.Memory.TotalGB) / 8.0 * 3.0
+	ramPower := c.capabilities.Memory.TotalGB / 8.0 * 3.0
 	power += ramPower
 
 	// Storage under load: 5-8W
